feat: refactor name dir for blog and mdx components
