+++ content/blog/part-batiment-ges-en-france.mdx
@@ -13,6 +13,10 @@ Le secteur du bâtiment représente aujourd'hui environ 18 % des émissions de g
 
 Pour atteindre la neutralité carbone d'ici 2050, la France doit drastiquement réduire les émissions de GES provenant des bâtiments, en les faisant passer de 75 millions de tonnes de CO2 (MtCO2) à 30 MtCO2 d'ici 2030.
 
+<div className="mb-8 relative w-full h-[400px]">
+  <Image src={"/img/blog/aaron-burden-bcn5fhJGtD8-unsplash.jpg"} alt="Image de décoration de l'article" fill className="rounded-lg" />
+</div>
+
 ### L’état actuel des bâtiments en France
 
 Actuellement près de la moitié des logements en France sont chauffés grâce à des énergies fossiles, comme le fioul ou le gaz. Si l'on ajoute l’électrique, 87 % des logements utilisent une source d'énergie non renouvelable pour le chauffage.
