+++ src/components/mdx/MdxDetails.tsx
